@@ -6,7 +6,6 @@
   "scripts": {
     "start": "nodemon ./bin/www",
     "watch:facemerge": "rimraf public &&cross-env ENV=dev PROJECT_NAME=facemerge node ./tools/webpack.watch.js",
-    "watch:voicemerge": "rimraf public &&cross-env ENV=dev PROJECT_NAME=voicemerge node ./tools/webpack.watch.js",
     "watch:viewport": "rimraf public &&cross-env ENV=dev PROJECT_NAME=viewport node ./tools/webpack.watch.js"
   },
   "dependencies": {
@@ -18,11 +17,11 @@
     "cross-env": "^5.1.4",
     "debug": "~2.6.9",
     "ejs": "~2.5.7",
-    "eruda": "^1.4.3",
     "express": "~4.15.5",
     "formidable": "^1.1.1",
     "morgan": "~1.9.0",
     "nodemon": "^1.17.1",
+    "postcss-px2rem": "^0.3.0",
     "qq-ai-sdk": "^1.0.3",
     "serve-favicon": "~2.4.5",
     "stream-buffers": "^3.0.1",
@@ -34,28 +33,23 @@
     "babel-loader": "^8.0.0-beta.0",
     "chalk": "^2.3.2",
     "css-loader": "^0.28.10",
-    "cssnano": "^3.10.0",
     "extract-text-webpack-plugin": "^3.0.2",
     "file-loader": "^1.1.11",
     "html-loader": "^0.5.5",
     "html-webpack-plugin": "^3.0.5",
     "inquirer": "^5.1.0",
     "jquery": "^3.3.1",
-    "postcss-cssnext": "^3.1.0",
-    "postcss-custom-properties": "^7.0.0",
-    "postcss-import": "^11.1.0",
+    "node-sass": "^4.8.3",
     "postcss-loader": "^2.1.3",
-    "postcss-px-to-viewport": "0.0.3",
     "postcss-safe-parser": "^3.0.1",
-    "postcss-write-svg": "github:jonathantneal/postcss-write-svg",
-    "postcss-viewport-units": "^0.1.4",
     "progress-bar-webpack-plugin": "^1.11.0",
     "raw-loader": "^0.5.1",
     "rimraf": "^2.6.2",
+    "sass-loader": "^6.0.7",
     "style-loader": "^0.20.2",
     "swiper": "^4.1.6",
     "uglifyjs-webpack-plugin": "^1.2.2",
-    "webpack": "^3.1.0",
+    "webpack": "^3.11.0",
     "webpack-cli": "^2.0.10",
     "webpack-dev-server": "^3.1.0",
     "y18n": "^4.0.0"
