@@ -1,5 +1,16 @@
-import '../assets/css/index.css';
-// import '../../common/js/flexible'
-const eruda = require('eruda');
+import '../../common/css/swiper.css';
+import '../assets/css/index.scss';
+import $ from 'jquery';
+import Swiper from 'swiper';
 
-eruda.init();
+$(document).ready(function(){
+    var mySwiper = new Swiper ('.swiper-container', {
+        direction: 'horizontal',
+        loop: true,
+
+        // 如果需要分页器
+        pagination: {
+            el: '.swiper-pagination',
+        }
+    })
+});
